requirements.txt
matplotlib
python-dotenv